+++ NEWS.md
@@ -3,6 +3,7 @@
 - Remove dependency on `sp`, `rgeos` and `rgdal` (#332)
 - That involved removal of the following functions:
   - [catchmentArea.R](https://github.com/ropensci/stplanr/blob/v0.8.5/R/catchmentArea.R)
+  - Some of the functionality from [linefuns.R](https://github.com/ropensci/stplanr/blob/v0.8.5/R/linefuns.R)
   
   - Browse the code base as of stplanr 0.8.5 here: https://github.com/ropensci/stplanr/blob/v0.8.5